requirements.txt
anyio==3.6.2
black==22.8.0
boto3==1.24.80
botocore==1.27.80
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
ecdsa==0.18.0
envs==1.4
fastapi==0.89.1
flake8==5.0.4
h11==0.14.0
httptools==0.5.0
idna==3.4
isort==5.10.1
jmespath==1.0.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
numpy==1.24.1
pandas==1.5.3
pathspec==0.10.1
platformdirs==2.5.2
pyasn1==0.4.8
pycodestyle==2.9.1
pycognito==2022.8.0
pycparser==2.21
pydantic==1.10.4
pyflakes==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-jose==3.3.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
types-jmespath==1.0.2
types-requests==2.28.11
types-urllib3==1.26.24
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
